fix return type of kvm_vcpu_get_hypercall_type

+++ include/vcpu.h
@@ -110,7 +110,7 @@ int kvm_vcpu_loop(struct kvm_vcpu *vcpu);
 */
 int kvm_vcpu_singlestep(struct kvm_vcpu *);
 
-int kvm_vcpu_get_hypercall_type(struct kvm_vm *, struct kvm_vcpu *);
+uint64_t kvm_vcpu_get_hypercall_type(struct kvm_vm *, struct kvm_vcpu *);
 
 int kvm_vcpu_had_page_fault(struct kvm_vcpu *);
